+++ src/fr/adrienbrault/idea/symfony2plugin/Symfony2ProjectComponent.java
@@ -143,7 +143,6 @@ public class Symfony2ProjectComponent implements ProjectComponent {
         }
 
         if(containerFiles.size() == 0) {
-            Symfony2ProjectComponent.getLogger().info("no custom container files add default");
             containerFiles.add(new ContainerFile(Settings.DEFAULT_CONTAINER_PATH));
         }
 
@@ -190,8 +189,30 @@ public class Symfony2ProjectComponent implements ProjectComponent {
 
     }
 
-    public static boolean isEnabled(Project project) {
-        return Settings.getInstance(project).pluginEnabled;
+    public static boolean isEnabled(@Nullable Project project) {
+        return project != null && Settings.getInstance(project).pluginEnabled;
+    }
+
+    /**
+     * If plugin is not enabled on first project start/indexing we will never get a filled
+     * index until a forced cache rebuild, we check also for vendor path
+     */
+    public static boolean isEnabledForIndex(Project project) {
+
+        if(Settings.getInstance(project).pluginEnabled) {
+            return true;
+        }
+
+        if(VfsUtil.findRelativeFile(project.getBaseDir(), "vendor", "symfony") != null) {
+            return true;
+        }
+
+        // drupal8; this should not really here
+        if(VfsUtil.findRelativeFile(project.getBaseDir(), "core", "vendor", "symfony") != null) {
+            return true;
+        }
+
+        return false;
     }
 
     public static boolean isEnabled(@Nullable PsiElement psiElement) {