+++ src/fr/adrienbrault/idea/symfony2plugin/stubs/indexes/AnnotationRoutesStubIndex.java
@@ -47,7 +47,7 @@ public class AnnotationRoutesStubIndex extends FileBasedIndexExtension<String, V
                 final Map<String, Void> map = new THashMap<String, Void>();
 
                 PsiFile psiFile = inputData.getPsiFile();
-                if(!Symfony2ProjectComponent.isEnabled(psiFile.getProject())) {
+                if(!Symfony2ProjectComponent.isEnabledForIndex(psiFile.getProject())) {
                     return map;
                 }