+++ src/fr/adrienbrault/idea/symfony2plugin/stubs/indexes/ServicesTagStubIndex.java
@@ -48,7 +48,7 @@ public class ServicesTagStubIndex extends FileBasedIndexExtension<String, String
                 Map<String, String[]> map = new THashMap<String, String[]>();
 
                 PsiFile psiFile = inputData.getPsiFile();
-                if(!Symfony2ProjectComponent.isEnabled(psiFile.getProject())) {
+                if(!Symfony2ProjectComponent.isEnabledForIndex(psiFile.getProject())) {
                     return map;
                 }